bootstrap/requirements.txt
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
gcloud==0.18.3
google-api-core==2.19.0
google-auth==2.30.0
google-cloud-core==2.4.1
google-cloud-iam==2.15.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.1
grpc-google-iam-v1==0.13.0
grpcio==1.64.1
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
oauth2client==4.1.3
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-terraform==0.10.1
requests==2.32.3
rsa==4.9
six==1.16.0
urllib3==2.2.1
